move db to seperate package
reasoning :
- no passing around of mdb connection struct
- the connection is only global within package scope which is fine imo and probably still improvable
- to use another db one only has to reimplement the interface
- overall nice capsulation

It also contains a query runner function now
The idea is to have a single point where queries are run for easier development of other features, i.e. our own concurrency management and whatnot
